store: split store functions into multiple files
