+++ api/store/publickey_store.go
@@ -1,0 +1,16 @@
+package store
+
+import (
+	"context"
+
+	"github.com/shellhub-io/shellhub/pkg/api/paginator"
+	"github.com/shellhub-io/shellhub/pkg/models"
+)
+
+type PublicKeyStore interface {
+	ListPublicKeys(ctx context.Context, pagination paginator.Query) ([]models.PublicKey, int, error)
+	GetPublicKey(ctx context.Context, fingerprint, tenant string) (*models.PublicKey, error)
+	CreatePublicKey(ctx context.Context, key *models.PublicKey) error
+	UpdatePublicKey(ctx context.Context, fingerprint, tenant string, key *models.PublicKeyUpdate) (*models.PublicKey, error)
+	DeletePublicKey(ctx context.Context, fingerprint, tenant string) error
+}
