+++ api/store/session_store.go
@@ -1,0 +1,20 @@
+package store
+
+import (
+	"context"
+
+	"github.com/shellhub-io/shellhub/pkg/api/paginator"
+	"github.com/shellhub-io/shellhub/pkg/models"
+)
+
+type SessionStore interface {
+	ListSessions(ctx context.Context, pagination paginator.Query) ([]models.Session, int, error)
+	GetSession(ctx context.Context, uid models.UID) (*models.Session, error)
+	CreateSession(ctx context.Context, session models.Session) (*models.Session, error)
+	SetSessionAuthenticated(ctx context.Context, uid models.UID, authenticated bool) error
+	KeepAliveSession(ctx context.Context, uid models.UID) error
+	DeactivateSession(ctx context.Context, uid models.UID) error
+	RecordSession(ctx context.Context, uid models.UID, record string, width, height int) error
+	UpdateUID(ctx context.Context, oldUID models.UID, newUID models.UID) error
+	GetRecord(ctx context.Context, uid models.UID) ([]models.RecordedSession, int, error)
+}
